# contracts/openapi/public-v1.yaml
openapi: 3.1.0
info:
  title: DengueSense Public API
  version: 1.0.0-draft.1
  summary: API công khai (qua gateway) cho dashboard DengueSense — phạm vi Đợt 1 (xem rủi ro).
  description: |
    **Nguồn sự thật của hợp đồng giữa `gateway` ↔ `dashboard`** (docs/09 §6.10).
    Code Go sinh từ file này (oapi-codegen), dashboard sinh type từ file này (openapi-typescript).

    Quy ước chung (docs/09 §6):
    - Trường JSON `snake_case`. Tháng dạng `"YYYY-MM"`, thời điểm RFC 3339 UTC.
    - Lỗi trả `application/problem+json` (RFC 9457) với `code` ổn định — danh mục ở `contracts/errors.md`.
    - Mọi response chứa số liệu mô hình có `meta` (provenance). Xác suất luôn đi cùng `base_rate`.
    - Client PHẢI bỏ qua trường lạ trong response (thêm trường tuỳ chọn không phải thay đổi phá vỡ).
    - Mọi `POST` tạo tài nguyên / có tác dụng phụ nhận header `Idempotency-Key`.

    Phạm vi bản này: đăng nhập, danh mục, quan sát, lượt dự báo, bản đồ rủi ro, giải thích, model card.
    Cảnh báo / hồ sơ / phân bổ / dự thảo / duyệt (Đợt 2–3) được thêm bằng PR hợp đồng riêng khi tới lượt.
servers:
  - url: /api/v1
    description: Cùng origin với dashboard (qua reverse proxy)
...
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Đăng nhập
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Thành công. Refresh token được đặt trong cookie `HttpOnly; Secure; SameSite=Strict; Path=/api/v1/auth`.
          headers:
            Set-Cookie:
              description: Cookie refresh token (HttpOnly).
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenResponse" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /auth/refresh:
    post:
      tags: [auth]
      operationId: refreshToken
      summary: Cấp access token mới bằng refresh token trong cookie (xoay vòng)
      security: []
      responses:
        "200":
          description: Access token mới; cookie refresh được xoay vòng.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: Đăng xuất (thu hồi refresh token trong cookie)
      description: |
        Công khai có chủ đích: phiên được xác định bằng cookie refresh, KHÔNG bằng access token — access token hết hạn
        (vd người dùng rời máy 15 phút) không được chặn đăng xuất, nếu không cookie vẫn sống và phiên tự khôi phục.
        Idempotent: không có cookie hoặc token đã thu hồi vẫn trả 204. `SameSite=Strict` chặn đăng xuất từ trang khác.
      security: []
      responses:
        "204":
          description: Đã đăng xuất; cookie refresh bị xoá.
          headers:
            Set-Cookie:
              description: Xoá cookie refresh (`Max-Age=0`).
              schema: { type: string }
        "429": { $ref: "#/components/responses/TooManyRequests" }
...
  /provinces:
    get:
      tags: [catalog]
      operationId: listProvinces
      summary: Danh mục 34 tỉnh (danh mục đóng, trả đủ — không phân trang)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Province" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /provinces/{province_id}:
    get:
      tags: [catalog]
      operationId: getProvince
      summary: Chi tiết một tỉnh
      parameters:
        - { $ref: "#/components/parameters/ProvinceId" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Province" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /geo/provinces:
    get:
      tags: [catalog]
      operationId: getProvinceGeometry
      summary: Ranh giới 34 tỉnh (GeoJSON, có phiên bản — bất biến theo `version`)
      parameters:
        - name: version
          in: query
          required: false
          description: Phiên bản ranh giới; mặc định là bản mới nhất. Response có `Cache-Control` dài khi chỉ định version.
          schema: { type: string, examples: ["2025-07"] }
      responses:
        "200":
          description: GeoJSON FeatureCollection; mỗi feature có `properties.province_id`.
          content:
            application/geo+json:
              schema: { $ref: "#/components/schemas/GeoJsonFeatureCollection" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /data-versions:
    get:
      tags: [catalog]
      operationId: listDataVersions
      summary: Các phiên bản dữ liệu (panel) đã công bố
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/DataVersion" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /observations:
    get:
      tags: [catalog]
      operationId: listObservations
      summary: Chuỗi quan sát theo tháng của một tỉnh
      parameters:
        - name: province_id
          in: query
          required: true
          schema: { $ref: "#/components/schemas/ProvinceId" }
        - name: from
          in: query
          required: false
          schema: { $ref: "#/components/schemas/YearMonth" }
        - name: to
          in: query
          required: false
          schema: { $ref: "#/components/schemas/YearMonth" }
        - name: data_version
          in: query
          required: false
          description: Mặc định là bản mới nhất.
          schema: { type: string, examples: ["v0.2.0"] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data_version, items]
                properties:
                  data_version: { type: string }
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Observation" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /forecast-runs:
    get:
      tags: [forecast]
      operationId: listForecastRuns
      summary: Danh sách lượt dự báo
      parameters:
        - name: mode
          in: query
          schema: { $ref: "#/components/schemas/RunMode" }
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/RunStatus" }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: OK (mới nhất trước)
          content:
            application/json:
              schema:
                type: object
                required: [items, next_cursor]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/ForecastRun" }
                  next_cursor:
                    type: [string, "null"]
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [forecast]
      operationId: createForecastRun
      summary: Tạo lượt dự báo (chạy nền). Vai trò tối thiểu `analyst`.
      description: |
        Luôn là job — mô hình không bao giờ chạy trong request đồng bộ (docs/09 §2 nguyên tắc 4).
        `backtest`: `origin_month` phải ≤ 2010-06 (giai đoạn có dữ liệu thật).
        `live_experimental`: kết quả luôn mang cờ `out_of_validated_period` và/hoặc `estimated_inputs`.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateForecastRunRequest" }
      responses:
        "202":
          description: Đã nhận, chạy nền.
          headers:
            Location:
              description: URL của job (`/api/v1/jobs/{job_id}`).
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /forecast-runs/{run_id}:
    get:
      tags: [forecast]
      operationId: getForecastRun
      summary: Chi tiết một lượt dự báo
      parameters:
        - { $ref: "#/components/parameters/RunId" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ForecastRun" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /risk-map:
    get:
      tags: [forecast]
      operationId: getRiskMap
      summary: Bản đồ rủi ro — dự báo 34 tỉnh cho một lượt và một tầm dự báo (BFF ghép nhiều service)
      description: |
        Một nguồn phụ lỗi (vd cảnh báo) → vẫn trả phần còn lại kèm `warnings[]`, không trả 500.
      parameters:
        - name: run_id
          in: query
          required: false
          description: Mặc định là lượt `completed` mới nhất.
          schema: { type: string, format: uuid }
        - name: horizon
          in: query
          required: true
          schema: { $ref: "#/components/schemas/Horizon" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RiskMap" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /provinces/{province_id}/forecasts:
    get:
      tags: [forecast]
      operationId: getProvinceForecasts
      summary: Dự báo một tỉnh cho đủ các tầm dự báo của một lượt
      parameters:
        - { $ref: "#/components/parameters/ProvinceId" }
        - name: run_id
          in: query
          required: false
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [meta, items]
                properties:
                  meta: { $ref: "#/components/schemas/Meta" }
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/ForecastItem" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /model-card:
    get:
      tags: [model]
      operationId: getModelCard
      summary: Model card rút gọn của phiên bản mô hình đang dùng
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModelCard" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token (EdDSA, 15 phút) lấy từ `/auth/login`.
...
  responses:
...
    NotFound:
      description: Không tìm thấy.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username: { type: string, minLength: 1, maxLength: 128 }
        password: { type: string, minLength: 1, maxLength: 256, writeOnly: true }
    TokenResponse:
      type: object
      required: [access_token, token_type, expires_in, user]
      properties:
        access_token: { type: string }
        token_type: { type: string, const: Bearer }
        expires_in: { type: integer, description: Số giây hiệu lực, examples: [900] }
        user: { $ref: "#/components/schemas/User" }
...